Generate unique names for rate limiter buckets (#137)
The current implementation of rate limiter bucket naming is not unique — it only uses the process ID (PID) to generate the folder name. As a result, when multiple PSN accounts are polled within the same process, they incorrectly share the same rate limiter. This causes the global rate limit to be applied across all accounts, instead of tracking limits per account as intended.

Co-authored-by: tr4nt0r <4445816+tr4nt0r@users.noreply.github.com>
Co-authored-by: Yoshikage Kira <8bit_yoshikage_kira@proton.me>
Reviewed-on: https://codeberg.org/YoshikageKira/psnawp/pulls/137
Co-authored-by: Manu <tr4nt0r@noreply.codeberg.org>
Co-committed-by: Manu <tr4nt0r@noreply.codeberg.org>